Holy forgotten commits

+++ engine/util/sphere.h
@@ -3,22 +3,24 @@
 #define VULPES_SPHERE_H
 
 #include "stdafx.h"
-#include "view_frustum.h"
-#include "AABB.h"
+#include "view_frustum.hpp"
+#include "AABB.hpp"
 namespace vulpes {
 
 	namespace util {
 
 		struct Sphere {
-			glm::dvec3 Origin;
-			double Radius;
+
+			glm::vec3 Origin;
+			float Radius;
+
 
 			Sphere() = default;
-			Sphere(const glm::dvec3& origin, const double& radius) : Origin(origin), Radius(radius) {};
+			Sphere(const glm::vec3& origin, const float& radius) : Origin(origin), Radius(radius) {};
 			
 			bool ConincidesWith(const Sphere& other) const {
 				// distance between spheres, sphere's intersect if this is less than sum of radii
-				double dist = glm::length(Origin - other.Origin);
+				float dist = glm::length(Origin - other.Origin);
 				return dist < Radius + other.Radius;
 			}
 
@@ -33,27 +35,27 @@ namespace vulpes {
 
 			bool CoincidesWith(const AABB& bounds) const {
 				// Get distance along each axis, sum it
-				double dmin = 0.0;
+				float dmin = 0.0;
 
 				if (Origin.x < bounds.Min.x) {
-					dmin += std::pow(Origin.x - bounds.Min.x, 2.0);
+					dmin += std::powf(Origin.x - bounds.Min.x, 2.0);
 				}
 				else if (Origin.x > bounds.Max.x) {
-					dmin += std::pow(Origin.x - bounds.Max.x, 2.0);
+					dmin += std::powf(Origin.x - bounds.Max.x, 2.0);
 				}
 
 				if (Origin.y < bounds.Min.y) {
-					dmin += std::pow(Origin.y - bounds.Min.y, 2.0);
+					dmin += std::powf(Origin.y - bounds.Min.y, 2.0);
 				}
 				else if (Origin.y > bounds.Max.y) {
-					dmin += std::pow(Origin.y - bounds.Max.y, 2.0);
+					dmin += std::powf(Origin.y - bounds.Max.y, 2.0);
 				}
 
 				if (Origin.z < bounds.Min.z) {
-					dmin += std::pow(Origin.z - bounds.Min.z, 2.0);
+					dmin += std::powf(Origin.z - bounds.Min.z, 2.0);
 				}
 				else if (Origin.z > bounds.Max.z) {
-					dmin += std::pow(Origin.z - bounds.Max.z, 2.0);
+					dmin += std::powf(Origin.z - bounds.Max.z, 2.0);
 				}
 
 				return (dmin <= std::pow(Radius, 2.0));
